+++ core/src/main/java/com/adobe/aem/commons/assetshare/configuration/impl/ConfigImpl.java
@@ -25,7 +25,6 @@ import com.adobe.aem.commons.assetshare.configuration.impl.selectors.AlwaysUseDe
 import com.adobe.aem.commons.assetshare.content.AssetModel;
 import com.adobe.aem.commons.assetshare.util.ForcedInheritanceValueMapWrapper;
 import com.day.cq.commons.inherit.HierarchyNodeInheritanceValueMap;
-import com.day.cq.dam.entitlement.api.EntitlementConstants;
 import com.day.cq.wcm.api.Page;
 import com.day.cq.wcm.api.PageManager;
 import com.day.cq.wcm.api.components.ComponentContext;
@@ -84,6 +83,8 @@ public class ConfigImpl implements Config {
     private static final String PN_ASSET_DETAILS_SELECTOR = "config/asset-details/selector";
     public static final String PN_PLACEHOLDER_ASSET_PATH = "config/asset-details/placeholderPath";
     public static final String PN_ASSET_REFERENCE_BY_ID = "config/asset-details/assetReferenceById";
+
+    @SuppressWarnings("CQRules:CQBP-71")
     public static final String DEFAULT_PLACEHOLDER_ASSET_PATH = "/apps/asset-share-commons/resources/placeholder.png";
 
     @Self
@@ -113,7 +114,6 @@ public class ConfigImpl implements Config {
 
     private String rootPath;
 
-    private Boolean dynamicMediaEnabled;
 
     @PostConstruct
     protected void init() {
@@ -233,20 +233,6 @@ public class ConfigImpl implements Config {
     @Override
     public String getRootPath() {
         return getRootPath(currentPage);
-    }
-
-    @Override
-    public boolean isDynamicMediaEnabled() {
-        if (dynamicMediaEnabled == null) {
-            if (features == null) {
-                dynamicMediaEnabled = Boolean.FALSE;
-            } else {
-                dynamicMediaEnabled = features.isEnabled(SCENE7_FEATURE_FLAG) ||
-                        features.isEnabled(EntitlementConstants.ASSETS_DYNAMICMEDIA_FEATURE_FLAG_PID);
-            }
-        }
-
-        return dynamicMediaEnabled;
     }
 
     /**